docs(openapi): create action to update API version in protobufs (#106)
Because

- API version should match latest release name in `instill-ai/core`.
- We rely on documentation and manual actions to keep these repos in
sync.

This commit

- Adds a GitHub action that creates a PR in the `protobufs` repo
modifying the
  API version.

## Notes 🗒️ 

I forked [core](https://github.com/jvallesm/instill-core) and
[protobufs](https://github.com/jvallesm/instill-protobufs) repos to test
this action (setting up my own API tokens).

### New release in `core`

![CleanShot 2024-01-19 at 12 01
05](https://github.com/instill-ai/core/assets/3977183/448c3283-0f62-41ca-b4b2-7cfcf4ef4797)
![CleanShot 2024-01-19 at 12 01
15](https://github.com/instill-ai/core/assets/3977183/8f0c7c2b-d3ba-43f7-ac71-70663dae3113)

### ⏭️ New PR in `protobufs`

See PR [here](https://github.com/jvallesm/instill-protobufs/pull/4).

![CleanShot 2024-01-19 at 12 03
08](https://github.com/instill-ai/core/assets/3977183/741471ea-c948-43c7-8991-89c3208e9f75)

**Note:** The PR and branch names omit intentionally the release name.
This makes the PR action to update the PR in `protobufs` if several
releases pile up.

### `TODO`: check OpenAPI file generation

The PR in the forked `protobufs` repo doesn't trigger the OpenAPI
generation workflow. As far as I understand we aren't using the
`GITHUB_TOKEN` variable so the workflow should be triggered. Perhaps
it's the config in my forked repo. Let's check in the next release, if
the files aren't generated we can open a PR and make generation part of
this action, adding a `buf` step and a `make openapi` command.